+++ rust/src/lib.rs
@@ -22,13 +22,13 @@ pub const FILE_DESCRIPTOR_SET: &[u8] = include_bytes!("proto_descriptor.bin");
 
 pub use crate::ics23::*;
 pub use api::{
-    iavl_spec, smt_spec, tendermint_spec, verify_batch_membership, verify_batch_non_membership,
-    verify_membership, verify_non_membership,
+    blake2b256_iavl_spec, blake3_iavl_spec, iavl_spec, membership_root, smt_spec, tendermint_spec,
+    verify_batch_membership, verify_batch_non_membership, verify_membership, verify_non_membership,
 };
 pub use compress::{compress, decompress, is_compressed};
 pub use helpers::{Hash, Result};
 pub use host_functions::HostFunctionsProvider;
-pub use verify::calculate_existence_root;
+pub use verify::{calculate_existence_root, verified_existence_root};
 
 #[cfg(feature = "host-functions")]
 pub use host_functions::host_functions_impl::HostFunctionsManager;